Added request headers to IRequestLogger, and moved interfaces to Core
