+++ src/SimpleRestServices/Client/HttpHeader.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace JSIStudios.SimpleRESTServices.Client
+{
+    [Serializable]
+    public class HttpHeader
+    {
+        public string Key { get; set; }
+
+        public string Value { get; set; }
+    }
+}
