+++ src/SimpleRestServices/Core/IRestService.cs
