+++ src/SimpleRestServices/Core/IUrlBuilder.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Collections.Generic;
+
+namespace JSIStudios.SimpleRESTServices.Core
+{
+    public interface IUrlBuilder
+    {
+        Uri Build(Uri baseUrl, Dictionary<string, string> queryStringParameters);
+    }
+}
