+++ src/SimpleRestServices/SimpleRestServices.csproj
@@ -46,8 +46,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Client\HttpHeader.cs" />
     <Compile Include="Client\HttpMethod.cs" />
-    <Compile Include="Client\IRestService.cs" />
+    <Compile Include="Core\IRestService.cs" />
+    <Compile Include="Core\IUrlBuilder.cs" />
+    <Compile Include="Core\IRequestLogger.cs" />
     <Compile Include="Client\Json\JsonRequestSettings.cs" />
     <Compile Include="Client\Json\JsonRestService.cs" />
     <Compile Include="Client\UrlBuilder.cs" />
@@ -55,7 +58,7 @@
     <Compile Include="Client\Response.cs" />
     <Compile Include="Client\WebResponseRetryLogic.cs" />
     <Compile Include="Core\ITextCleaner.cs" />
-    <Compile Include="IRetryLogic.cs" />
+    <Compile Include="Core\IRetryLogic.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Server\FullHtmlTagCleaner.cs" />
     <Compile Include="Server\HtmlTagReplacerAndCleaner.cs" />
